google_adk_agents/requirements.txt
# Google ADK Requirements for Wellspring Em Dash Agent Framework
google-adk>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
asyncio-tools>=0.1.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0

# Existing Wellspring dependencies
sqlite3-tools>=0.1.0
typing-extensions>=4.0.0

# Optional but recommended for production
uvloop>=0.17.0  # For better async performance